Update UI in word_search_view.py
The word search configuration window now opens after a button on the root window is clicked instead of appearing at the same time as the root but being behind the root window. The 'open_editor' method was added to accomplish this.

The 'Solve Word Search' button on the root window has been re-written to perform two actions: 1) Open the word search configuration window, and 2) Invoke the word search solver algorithm. This new action button will be enabled and disabled depending on the stage of the word search program to prevent accidental presses.